+++ packages/public-api/src/integration.test.ts
@@ -1,8 +1,13 @@
 import { describe, expect, it } from 'vitest'
 
+import type { Asset, AssetsListResponse } from './routes/assets/types'
+import type { ChainsListResponse } from './routes/chains/types'
+import type { QuoteResponse } from './routes/quote/types'
+import type { RateResponse } from './routes/rates/types'
+
 const API_URL = process.env.API_URL ?? `http://localhost:${process.env.PORT ?? '3001'}`
 
-const TEST_PARTNER_CODE = 'test-partner'
+const VULTISIG_PARTNER_CODE = 'vultisig'
 
 const ASSET_IDS = {
   ETH: 'eip155:1/slip44:60',
@@ -27,7 +32,7 @@ describe('health', () => {
   it('returns ok', async () => {
     const res = await fetch(`${API_URL}/health`)
     expect(res.ok).toBe(true)
-    const data = await res.json()
+    const data = (await res.json()) as { status: string }
     expect(data.status).toBe('ok')
   })
 })
@@ -36,7 +41,7 @@ describe('/v1/chains', () => {
   it('returns a non-empty list of chains', async () => {
     const res = await fetch(`${API_URL}/v1/chains`)
     expect(res.ok).toBe(true)
-    const data = await res.json()
+    const data = (await res.json()) as ChainsListResponse
     expect(Array.isArray(data.chains)).toBe(true)
     expect(data.chains.length).toBeGreaterThan(0)
     const [first] = data.chains
@@ -48,7 +53,7 @@ describe('/v1/assets', () => {
   it('returns a non-empty list of assets', async () => {
     const res = await fetch(`${API_URL}/v1/assets?limit=10`)
     expect(res.ok).toBe(true)
-    const data = await res.json()
+    const data = (await res.json()) as AssetsListResponse
     expect(Array.isArray(data.assets)).toBe(true)
     expect(data.assets.length).toBeGreaterThan(0)
   })
@@ -56,7 +61,7 @@ describe('/v1/assets', () => {
   it('returns a single asset by id', async () => {
     const res = await fetch(`${API_URL}/v1/assets/${encodeURIComponent(ASSET_IDS.ETH)}`)
     expect(res.ok).toBe(true)
-    const asset = await res.json()
+    const asset = (await res.json()) as Asset
     expect(asset).toMatchObject({ chainId: expect.any(String), symbol: expect.any(String) })
   })
 })
@@ -76,14 +81,13 @@ describe('/v1/swap/rates', () => {
     const { sellAssetId, buyAssetId, sellAmountCryptoBaseUnit } = TEST_PAIRS.evmSameChain
     const params = new URLSearchParams({ sellAssetId, buyAssetId, sellAmountCryptoBaseUnit })
     const res = await fetch(`${API_URL}/v1/swap/rates?${params}`, {
-      headers: { 'X-Partner-Code': TEST_PARTNER_CODE },
+      headers: { 'X-Partner-Code': VULTISIG_PARTNER_CODE },
     })
     expect(res.ok).toBe(true)
-    const data = await res.json()
+    const data = (await res.json()) as RateResponse
     expect(Array.isArray(data.rates)).toBe(true)
     const validRates = data.rates.filter(
-      (r: { error?: unknown; buyAmountCryptoBaseUnit?: string }) =>
-        !r.error && r.buyAmountCryptoBaseUnit && r.buyAmountCryptoBaseUnit !== '0',
+      r => !r.error && r.buyAmountCryptoBaseUnit && r.buyAmountCryptoBaseUnit !== '0',
     )
     expect(validRates.length).toBeGreaterThan(0)
   })
@@ -92,16 +96,38 @@ describe('/v1/swap/rates', () => {
     const { sellAssetId, buyAssetId, sellAmountCryptoBaseUnit } = TEST_PAIRS.crossChain
     const params = new URLSearchParams({ sellAssetId, buyAssetId, sellAmountCryptoBaseUnit })
     const res = await fetch(`${API_URL}/v1/swap/rates?${params}`, {
-      headers: { 'X-Partner-Code': TEST_PARTNER_CODE },
+      headers: { 'X-Partner-Code': VULTISIG_PARTNER_CODE },
     })
     expect(res.ok).toBe(true)
-    const data = await res.json()
+    const data = (await res.json()) as RateResponse
     expect(Array.isArray(data.rates)).toBe(true)
     const validRates = data.rates.filter(
-      (r: { error?: unknown; buyAmountCryptoBaseUnit?: string }) =>
-        !r.error && r.buyAmountCryptoBaseUnit && r.buyAmountCryptoBaseUnit !== '0',
+      r => !r.error && r.buyAmountCryptoBaseUnit && r.buyAmountCryptoBaseUnit !== '0',
     )
     expect(validRates.length).toBeGreaterThan(0)
+  })
+
+  it('uses default affiliate bps when no partner code is provided', async () => {
+    const { sellAssetId, buyAssetId, sellAmountCryptoBaseUnit } = TEST_PAIRS.evmSameChain
+    const params = new URLSearchParams({ sellAssetId, buyAssetId, sellAmountCryptoBaseUnit })
+    const res = await fetch(`${API_URL}/v1/swap/rates?${params}`)
+    expect(res.ok).toBe(true)
+    const data = (await res.json()) as RateResponse
+    const [first] = data.rates
+    expect(first.affiliateBps).toBe('60')
+  })
+
+  it('uses resolved affiliate bps when a valid partner code is provided', async () => {
+    const { sellAssetId, buyAssetId, sellAmountCryptoBaseUnit } = TEST_PAIRS.evmSameChain
+    const params = new URLSearchParams({ sellAssetId, buyAssetId, sellAmountCryptoBaseUnit })
+    const res = await fetch(`${API_URL}/v1/swap/rates?${params}`, {
+      headers: { 'X-Partner-Code': VULTISIG_PARTNER_CODE },
+    })
+    expect(res.ok).toBe(true)
+    const data = (await res.json()) as RateResponse
+    const [first] = data.rates
+    expect(typeof first.affiliateBps).toBe('string')
+    expect(Number(first.affiliateBps)).toBeGreaterThan(0)
   })
 })
 
@@ -119,5 +145,7 @@ describe('/v1/swap/quote', () => {
       }),
     })
     expect(res.status).not.toBe(401)
+    const data = (await res.json()) as QuoteResponse
+    expect(data).toMatchObject({ quoteId: expect.any(String), swapperName: expect.any(String) })
   })
 })
