+++ packages/public-api/src/registry.ts
@@ -1,0 +1,5 @@
+import './setupZod'
+
+import { OpenAPIRegistry } from '@asteasolutions/zod-to-openapi'
+
+export const registry = new OpenAPIRegistry()
