+++ packages/public-api/src/routes/affiliate/index.ts
@@ -1,0 +1,1 @@
+export * from './getAffiliateStats'
