+++ packages/public-api/src/routes/assets/index.ts
@@ -1,0 +1,3 @@
+export * from './getAssets'
+export * from './getAssetById'
+export * from './getAssetCount'
