+++ packages/public-api/src/routes/chains/index.ts
@@ -1,0 +1,2 @@
+export * from './getChains'
+export * from './getChainCount'
