+++ packages/public-api/src/routes/chains/utils.ts
@@ -1,9 +1,8 @@
 import { fromChainId } from '@shapeshiftoss/caip'
 import { KnownChainIds } from '@shapeshiftoss/types'
 import { getBaseAsset } from '@shapeshiftoss/utils'
-import type { Request, Response } from 'express'
 
-import type { Chain, ChainsResponse, ChainType, ErrorResponse } from '../types'
+import type { Chain, ChainType } from './types'
 
 const chainNamespaceToType: Record<string, ChainType> = {
   eip155: 'evm',
@@ -56,35 +55,9 @@ const buildChainList = (): Chain[] => {
 
 let cachedChains: Chain[] | null = null
 
-const getChainList = (): Chain[] => {
+export const getChainList = (): Chain[] => {
   if (!cachedChains) {
     cachedChains = buildChainList()
   }
   return cachedChains
-}
-
-export const getChains = (_req: Request, res: Response): void => {
-  try {
-    const chains = getChainList()
-
-    const response: ChainsResponse = {
-      chains,
-      timestamp: Date.now(),
-    }
-
-    res.json(response)
-  } catch (error) {
-    console.error('Error in getChains:', error)
-    res.status(500).json({ error: 'Internal server error' } as ErrorResponse)
-  }
-}
-
-export const getChainCount = (_req: Request, res: Response): void => {
-  try {
-    const count = getChainList().length
-    res.json({ count, timestamp: Date.now() })
-  } catch (error) {
-    console.error('Error in getChainCount:', error)
-    res.status(500).json({ error: 'Internal server error' } as ErrorResponse)
-  }
 }
