+++ packages/public-api/src/routes/docs.ts
@@ -1,7 +1,34 @@
+import { OpenApiGeneratorV3 } from '@asteasolutions/zod-to-openapi'
 import { apiReference } from '@scalar/express-api-reference'
 import express from 'express'
+import { readFileSync } from 'fs'
+import { join } from 'path'
 
-import { generateOpenApiDocument } from '../docs/openapi'
+import { registry } from '../registry'
+
+const readDoc = (filename: string) => readFileSync(join(__dirname, '../docs', filename), 'utf-8')
+
+const generateOpenApiDocument = () => {
+  const generator = new OpenApiGeneratorV3(registry.definitions)
+
+  const doc = generator.generateDocument({
+    openapi: '3.0.0',
+    info: {
+      version: '1.0.0',
+      title: 'ShapeShift Public API',
+      description: readDoc('introduction.md'),
+    },
+    servers: [{ url: 'https://api.shapeshift.com' }, { url: 'http://localhost:3001' }],
+  })
+
+  doc.tags = [
+    { name: 'Swap Widget SDK', description: readDoc('swap-widget-sdk.md') },
+    { name: 'REST API Guide', description: readDoc('rest-api-guide.md') },
+    ...(doc.tags ?? []),
+  ]
+
+  return doc
+}
 
 const router = express.Router()
