+++ packages/public-api/src/routes/quote/index.ts
@@ -1,0 +1,1 @@
+export * from './getQuote'
