+++ packages/public-api/src/routes/rates/index.ts
@@ -1,0 +1,1 @@
+export * from './getRates'
