+++ packages/public-api/src/routes/status/constants.ts
@@ -1,0 +1,1 @@
+export const STATUS_TIMEOUT_MS = 10_000
