+++ packages/public-api/src/routes/status/getSwapStatus.ts
@@ -1,87 +1,43 @@
 import type { Request, Response } from 'express'
-import { z } from 'zod'
 
-import { getAsset } from '../assets'
-import { SWAP_SERVICE_BASE_URL } from '../config'
-import { quoteStore } from '../lib/quoteStore'
-import type { ErrorResponse } from '../types'
+import { SWAP_SERVICE_BASE_URL } from '../../config'
+import { quoteStore } from '../../lib/quoteStore'
+import { registry } from '../../registry'
+import type { ErrorResponse } from '../../types'
+import { PartnerCodeHeaderSchema } from '../../types'
+import { STATUS_TIMEOUT_MS } from './constants'
+import type { SwapServiceStatus, SwapStatusResponse } from './types'
+import { StatusRequestSchema, SwapStatusResponseSchema } from './types'
+import { registerSwapInService } from './utils'
 
-const STATUS_TIMEOUT_MS = 10_000
-
-export const StatusRequestSchema = z.object({
-  quoteId: z.string().uuid(),
-  txHash: z.string().min(1).max(128).optional(),
+registry.registerPath({
+  method: 'get',
+  path: '/v1/swap/status',
+  operationId: 'getSwapStatus',
+  summary: 'Get swap status',
+  description:
+    'Look up the current status of a swap by its quote ID. Pass txHash on the first call after broadcasting to bind it to the quote and start tracking. Subsequent calls can omit txHash.',
+  tags: ['Swaps'],
+  request: {
+    headers: PartnerCodeHeaderSchema,
+    query: StatusRequestSchema,
+  },
+  responses: {
+    200: {
+      description: 'Swap status',
+      content: { 'application/json': { schema: SwapStatusResponseSchema } },
+    },
+    400: {
+      description: 'Invalid request parameters',
+    },
+    404: {
+      description: 'Quote not found or expired',
+    },
+    409: {
+      description: 'Transaction hash mismatch',
+    },
+  },
 })
-
-const toHumanAmount = (baseUnit: string, precision: number): string => {
-  if (precision === 0) return baseUnit
-  const padded = baseUnit.padStart(precision + 1, '0')
-  return `${padded.slice(0, -precision)}.${padded.slice(-precision)}`
-}
-
-const buildSwapRegistrationBody = (storedQuote: ReturnType<typeof quoteStore.get> & object) => {
-  const sellAsset = getAsset(storedQuote.sellAssetId)
-  const buyAsset = getAsset(storedQuote.buyAssetId)
-  if (!sellAsset || !buyAsset) return undefined
-
-  return {
-    body: JSON.stringify({
-      swapId: storedQuote.quoteId,
-      sellAsset,
-      buyAsset,
-      sellAmountCryptoBaseUnit: storedQuote.sellAmountCryptoBaseUnit,
-      expectedBuyAmountCryptoBaseUnit: storedQuote.buyAmountAfterFeesCryptoBaseUnit,
-      sellAmountCryptoPrecision: toHumanAmount(
-        storedQuote.sellAmountCryptoBaseUnit,
-        sellAsset.precision,
-      ),
-      expectedBuyAmountCryptoPrecision: toHumanAmount(
-        storedQuote.buyAmountAfterFeesCryptoBaseUnit,
-        buyAsset.precision,
-      ),
-      sellTxHash: storedQuote.txHash,
-      source: storedQuote.swapperName,
-      swapperName: storedQuote.swapperName,
-      sellAccountId: storedQuote.sendAddress,
-      receiveAddress: storedQuote.receiveAddress,
-      affiliateAddress: storedQuote.affiliateAddress,
-      affiliateBps: storedQuote.affiliateBps,
-      origin: 'api',
-      metadata: storedQuote.metadata,
-    }),
-    sellAsset,
-    buyAsset,
-  }
-}
-
-const registerSwapInService = async (
-  storedQuote: ReturnType<typeof quoteStore.get> & object,
-): Promise<boolean> => {
-  const registration = buildSwapRegistrationBody(storedQuote)
-  if (!registration) return false
-
-  const controller = new AbortController()
-  const timeout = setTimeout(() => controller.abort(), STATUS_TIMEOUT_MS)
-  try {
-    const postResponse = await fetch(`${SWAP_SERVICE_BASE_URL}/swaps`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      signal: controller.signal,
-      body: registration.body,
-    })
-    if (!postResponse.ok) {
-      const errorBody = await postResponse.text()
-      console.error(`swap-service POST failed (${postResponse.status}):`, errorBody)
-      return false
-    }
-    return true
-  } catch (err) {
-    console.error('Failed to register swap in swap-service:', err)
-    return false
-  } finally {
-    clearTimeout(timeout)
-  }
-}
 
 export const getSwapStatus = async (req: Request, res: Response): Promise<void> => {
   try {
@@ -142,7 +98,7 @@ export const getSwapStatus = async (req: Request, res: Response): Promise<void> 
       await registerSwapInService(storedQuote)
     }
 
-    let swapServiceStatus: Record<string, unknown> | null = null
+    let swapServiceStatus: SwapServiceStatus | null = null
     if (storedQuote.txHash) {
       const getController = new AbortController()
       const getTimeout = setTimeout(() => getController.abort(), STATUS_TIMEOUT_MS)
@@ -151,7 +107,7 @@ export const getSwapStatus = async (req: Request, res: Response): Promise<void> 
           signal: getController.signal,
         })
         if (swapResponse.ok) {
-          swapServiceStatus = (await swapResponse.json()) as Record<string, unknown>
+          swapServiceStatus = (await swapResponse.json()) as SwapServiceStatus
         } else if (swapResponse.status === 404) {
           await registerSwapInService(storedQuote)
         }
@@ -174,7 +130,7 @@ export const getSwapStatus = async (req: Request, res: Response): Promise<void> 
       quoteStore.set(quoteId, storedQuote)
     }
 
-    const response: Record<string, unknown> = {
+    const response: SwapStatusResponse = {
       quoteId,
       txHash: storedQuote.txHash,
       status,
